fix(lint): keep the events package within its file-count baseline
The nil-receiver coverage tests were added as a new file, which took
pkg/services/recordings/internal/events from 19 to 20 files and tripped
pkg-file-count ("package grew beyond baseline: recorded 19, now 20"). That
was the sole failing target in the Backend Lint ratchet verdict.

Fold the two tests into the existing event_history_test.go rather than
raising the recorded baseline: the baseline is a deletion-only ratchet and
bumping it to accommodate new tests spends budget the package does not need
to spend. File count returns to 19 and pkg-file-count holds.
